separated calculation from main
